+++ package.json
@@ -30,6 +30,7 @@
     "discord-api-types": "^0.37.104",
     "discord.js": "^14.16.3",
     "dotenv": "^16.4.5",
+    "reflect-metadata": "^0.2.2",
     "typescript": "^5.6.3"
   },
   "devDependencies": {
