+++ src/commands/ping.ts
@@ -1,9 +1,15 @@
 import { SlashCommandBuilder } from "discord.js";
+import {AddExecute, CommandData, InteractionHandler} from "../util/interaction-handler";
 
-module.exports = {
-  data: new SlashCommandBuilder()
+@CommandData(
+  new SlashCommandBuilder()
     .setName("ping")
-    .setDescription("현재 서버의 응답 속도를 확인합니다."),
+    .setDescription("현재 서버의 응답 속도를 확인합니다.")
+    .toJSON()
+)
+@InteractionHandler()
+export default class PingCommand {
+  @AddExecute("ping")
   async execute(interaction: any) {
     const sent = await interaction.reply({
       content: "Pong!",
@@ -11,5 +17,5 @@ module.exports = {
     });
     const timeTaken = sent.createdTimestamp - interaction.createdTimestamp;
     await interaction.followUp(`응답 속도: ${timeTaken}ms`);
-  },
-};
+  }
+}
